refactor: centralize model architectures and drop catalyst
