+++ requirements-core.in
@@ -24,7 +24,6 @@ gymnasium>=0.29.1
 mlflow>=3.2.0
 pytorch-lightning>=2.5.2
 transformers==4.56.1
-catalyst==21.5
 fastapi-csrf-protect>=1.0.6
 flask>=3.0.2
 pybit>=5.11.0
