+++ requirements.out
@@ -58,7 +58,6 @@ cachetools==5.5.2
     #   google-auth
     #   mlflow-skinny
     #   mlflow-tracing
-catalyst==21.5
     # via -r requirements.txt
 ccxt==4.5.3
     # via -r requirements.txt
@@ -330,7 +329,6 @@ numba==0.61.2
 numpy==2.2.6
     # via
     #   -r requirements.txt
-    #   catalyst
     #   contourpy
     #   gymnasium
     #   imbalanced-learn
@@ -505,7 +503,6 @@ pytz==2025.2
 pyyaml==6.0.2
     # via
     #   -r requirements.txt
-    #   catalyst
     #   huggingface-hub
     #   mlflow-skinny
     #   optuna
@@ -610,7 +607,6 @@ tenacity==9.1.2
 tensorboardx==2.6.4
     # via
     #   -r requirements.txt
-    #   catalyst
 threadpoolctl==3.6.0
     # via
     #   -r requirements.txt
@@ -625,7 +621,6 @@ tomli==2.2.1
 torch==2.8.0
     # via
     #   -r requirements.txt
-    #   catalyst
     #   pytorch-lightning
     #   stable-baselines3
     #   torchmetrics
@@ -636,7 +631,6 @@ torchmetrics==1.8.1
 tqdm==4.67.1
     # via
     #   -r requirements.txt
-    #   catalyst
     #   huggingface-hub
     #   openai
     #   optuna
